[explorer/cpu_cores] Try to improve #cores detection
When running under KVM lparstat psize reports the number of configured CPU
threads (mode=Capped).
When running under PowerVM in Uncapped mode psize reports the number of CPU
cores of the whole system, not just those assigned to the LPAR.

This commit tries to clean this up making a more sophisticated calculation.

"Maximum Virtual CPUs" is the number of cores which can be unfolded while
"Online Virtual CPUs" is the number of cores currently utilised by the LPAR.

$smt is the averaged SMT count for all processors. I don't know if different
sockets/cores can have a different SMT count, but for some reason smtctl(1) lists
the SMT threads for every socket. *shrug*